+++ packages/pf3-component-mapper/src/form-fields/select/select-helper.js
